@@ -13,10 +13,14 @@ namespace TheThroneOfGames.API.Controllers
     public class UsuarioController : ControllerBase
     {
         private readonly IUsuarioService _userService;
+        private readonly Services.AuthenticationService _authService;
+        private readonly TheThroneOfGames.Infrastructure.ExternalServices.EmailService _emailService;
 
-        public UsuarioController(IUsuarioService userService)
+        public UsuarioController(IUsuarioService userService, Services.AuthenticationService authService, TheThroneOfGames.Infrastructure.ExternalServices.EmailService emailService)
         {
             _userService = userService;
+            _authService = authService;
+            _emailService = emailService;
         }
 
         // GET: api/<UsuarioController>
@@ -40,7 +44,13 @@ namespace TheThroneOfGames.API.Controllers
         [AllowAnonymous]
         public IActionResult Post([FromBody] UserDTO value)
         {
-            return Ok("Usuário registrado com sucesso!");
+            // Forward to service for secure registration and send activation e-mail
+            var activationToken = _userService.PreRegisterUserAsync(value.Email, value.Name, value.Password).GetAwaiter().GetResult();
+            var activationLink = $"{Request.Scheme}://{Request.Host}/api/Usuario/activate?activationToken={activationToken}";
+            var subject = "Ativação de conta - TheThroneOfGames";
+            var body = $"Olá {value.Name},\n\nPor favor ative sua conta clicando no link abaixo:\n{activationLink}\n\nSe você não solicitou esse e-mail, ignore.";
+            _emailService.SendEmailAsync(value.Email, subject, body).GetAwaiter().GetResult();
+            return Ok("Usuário registrado com sucesso! E-mail de ativação enviado.");
         }
 
         // PUT api/<UsuarioController>/5
@@ -64,9 +74,27 @@ namespace TheThroneOfGames.API.Controllers
             // Validação do formato do e-mail
             if (!Regex.IsMatch(userDto.Email, @"^[^@\s]+@[^@\s]+\.[^@\s]+$"))
                 return BadRequest("E-mail inválido.");
+            // Validação de senha
+            try
+            {
+                var activationToken = await _userService.PreRegisterUserAsync(userDto.Email, userDto.Name, userDto.Password);
 
-            // Enviar e-mail de ativação
-            //await _emailService.SendEmailAsync(userDto.Email, "Ativação de Conta", "Clique no link para ativar sua conta.");
+                // Compose activation link
+                var activationLink = $"{Request.Scheme}://{Request.Host}/api/Usuario/activate?activationToken={activationToken}";
+
+                // Send activation e-mail (development: writes to Outbox)
+                var subject = "Ativação de conta - TheThroneOfGames";
+                var body = $"Olá {userDto.Name},\n\nPor favor ative sua conta clicando no link abaixo:\n{activationLink}\n\nSe você não solicitou esse e-mail, ignore.";
+
+                await _emailService.SendEmailAsync(userDto.Email, subject, body);
+            }
+            catch (ArgumentException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+
+            // Enviar e-mail de ativação (placeholder)
+            // await _emailService.SendEmailAsync(userDto.Email, "Ativação de Conta", "Clique no link para ativar sua conta.");
 
             return Ok("E-mail de ativação enviado.");
         }
@@ -79,6 +107,20 @@ namespace TheThroneOfGames.API.Controllers
             await _userService.ActivateUserAsync(activationToken);
 
             return Ok("Usuário ativado com sucesso.");
+        }
+
+        [HttpPost("login")]
+        [AllowAnonymous]
+        public async Task<IActionResult> Login([FromBody] LoginDTO loginDto)
+        {
+            if (string.IsNullOrWhiteSpace(loginDto.Email) || string.IsNullOrWhiteSpace(loginDto.Password))
+                return BadRequest("E-mail e senha são obrigatórios.");
+
+            var token = await _authService.AuthenticateAsync(loginDto.Email, loginDto.Password);
+            if (token == null)
+                return Unauthorized("Credenciais inválidas ou conta não ativada.");
+
+            return Ok(new { token });
         }
     }
 }
