@@ -5,15 +5,16 @@ namespace TheThroneOfGames.API.Models.DTO
     {
         public string Name { get; set; }
         public string Email { get; set; }
-        public string PasswordHash { get; set; }
+        // Accept raw password from client; service will hash
+        public string Password { get; set; }
         public string Role { get; set; }
 
         public UserDTO() { }
-        public UserDTO(string name, string email, string passwordHash, string role)
+        public UserDTO(string name, string email, string password, string role)
         {
             Name = name;
             Email = email;
-            PasswordHash = passwordHash;
+            Password = password;
             Role = role;
         }
     }
