@@ -7,13 +7,22 @@ using TheThroneOfGames.Application;
 using TheThroneOfGames.Domain.Interfaces;
 using TheThroneOfGames.Infrastructure.Repository;
 using Microsoft.AspNetCore.Builder;
+using TheThroneOfGames.Infrastructure.Persistence;
+using Microsoft.EntityFrameworkCore;
 
 
 var builder = WebApplication.CreateBuilder(args);
 // Register application services
-#region Inje��o de dependencias
-builder.Services.AddScoped<IUsuarioService, UsuarioService>();
-builder.Services.AddScoped<IUsuarioRepository, UsuarioRepository>();
+builder.Services.AddDbContext<MainDbContext>(options =>
+    options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
+
+#region Injeção de dependencias
+// Add application services
+builder.Services.AddApplicationServices();
+
+// Authentication & email
+builder.Services.AddScoped<TheThroneOfGames.API.Services.AuthenticationService>();
+builder.Services.AddScoped<TheThroneOfGames.Infrastructure.ExternalServices.EmailService>();
 #endregion
 
 
@@ -44,30 +53,54 @@ builder.Services.AddControllers(); // Adiciona suporte a controllers
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen(options =>
 {
-    // Opcional: Personalizar a documenta��o do Swagger
-    // Voc� pode adicionar informa��es sobre sua API aqui.
+    // Opcional: Personalizar a documenta��o do Swagger
+    // Voc� pode adicionar informa��es sobre sua API aqui.
     options.SwaggerDoc("v1", new Microsoft.OpenApi.Models.OpenApiInfo
     {
-        Title = "Minha Super Minimal API", // T�tulo da sua API
-        Version = "v1", // Vers�o da API
-        Description = "Uma API de exemplo para gerenciar produtos e clientes.", // Descri��o
-        TermsOfService = new Uri("https://example.com/terms"), // Termos de Servi�o (opcional)
+        Title = "Minha Super Minimal API", // T�tulo da sua API
+        Version = "v1", // Vers�o da API
+        Description = "Uma API de exemplo para gerenciar produtos e clientes.", // Descri��o
+        TermsOfService = new Uri("https://example.com/terms"), // Termos de Servi�o (opcional)
         Contact = new Microsoft.OpenApi.Models.OpenApiContact // Contato (opcional)
         {
             Name = "Seu Nome",
             Email = "seu.email@example.com"
         },
-        License = new Microsoft.OpenApi.Models.OpenApiLicense // Licen�a (opcional)
+        License = new Microsoft.OpenApi.Models.OpenApiLicense // Licen�a (opcional)
         {
-            Name = "Licen�a MIT",
+            Name = "Licen�a MIT",
             Url = new Uri("https://opensource.org/licenses/MIT")
         }
 
     });
+    // Add JWT Bearer authentication to Swagger (Authorize button)
+    options.AddSecurityDefinition("Bearer", new Microsoft.OpenApi.Models.OpenApiSecurityScheme
+    {
+        Description = "Insira o token JWT no campo: 'Bearer {token}'\n\nExemplo: 'Bearer eyJhbGci...'.",
+        Name = "Authorization",
+        In = Microsoft.OpenApi.Models.ParameterLocation.Header,
+        Type = Microsoft.OpenApi.Models.SecuritySchemeType.Http,
+        Scheme = "bearer",
+        BearerFormat = "JWT"
+    });
+
+    options.AddSecurityRequirement(new Microsoft.OpenApi.Models.OpenApiSecurityRequirement
+    {
+        {
+            new Microsoft.OpenApi.Models.OpenApiSecurityScheme
+            {
+                Reference = new Microsoft.OpenApi.Models.OpenApiReference
+                {
+                    Type = Microsoft.OpenApi.Models.ReferenceType.SecurityScheme,
+                    Id = "Bearer"
+                }
+            },
+            new string[] { }
+        }
+    });
 });
 
 var app = builder.Build();
-
 // Configure o pipeline HTTP para usar Swagger
 if (app.Environment.IsDevelopment())
 {
@@ -77,11 +110,16 @@ if (app.Environment.IsDevelopment())
 
 app.UseHttpsRedirection();
 
+// Global exception handling middleware
+app.UseMiddleware<TheThroneOfGames.API.Middleware.ExceptionMiddleware>();
+
 app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapControllers(); // Garante que os controllers sejam mapeados
 
-app.Run(); // Mant�m a aplica��o rodando
+app.Run(); // Mantém a aplicação rodando
+
+public partial class Program { }
 
 
